refactor: move social types to its own file
