+++ src/app/components/Social/Social.tsx
@@ -1,41 +1,7 @@
 import cn from 'classnames'
-import { IconProp } from '@fortawesome/fontawesome-svg-core'
-import {
-  faGithub,
-  faInstagram,
-  faLinkedinIn,
-  faTwitch,
-} from '@fortawesome/free-brands-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
-type SocialMediaType = {
-  href: string
-  icon: IconProp
-  iconClass?: string
-}
-
-const socialMedias: SocialMediaType[] = [
-  {
-    href: 'https://www.linkedin.com/in/lucas-felicio-adriano/',
-    icon: faLinkedinIn,
-    iconClass: 'text-[#0072b1]',
-  },
-  {
-    href: 'https://www.instagram.com/lucas.adrianoo/',
-    icon: faInstagram,
-    iconClass: 'text-[#DD2A7B]',
-  },
-  {
-    href: 'https://github.com/lucasadrianof/',
-    icon: faGithub,
-    iconClass: 'text-white',
-  },
-  {
-    href: 'https://twitch.tv/the_branko',
-    icon: faTwitch,
-    iconClass: 'text-[#6441A4]',
-  },
-]
+import socialMedias from './data'
 
 export default function Social() {
   return (
@@ -44,7 +10,7 @@ export default function Social() {
         <a key={key} href={href} target="_blank" rel="noreferrer">
           <span
             className={cn(
-              'w-10 h-10 rounded-lg hover:bg-gradient-to-r from-[#FA5252] to-[#DD2476]  flex justify-center items-center  hover:text-white transition-colors duration-300  ease-in-out  bg-[#212425]',
+              'w-10 h-10 rounded-lg hover:bg-gradient-to-r from-[#FA5252] to-[#DD2476] flex justify-center items-center  hover:text-white transition-colors duration-300  ease-in-out  bg-[#212425]',
               iconClass
             )}
           >
